requirements.txt
streamlit
openpyxl
pygwalker
pandas
matplotlib.pyplot
seaborn
datetime
sys
eda_tab
eda_dashboard_tab
statsmodels.api
streamlit
pandas
numpy
matplotlib.pyplot
math
sklearn.linear_model
LinearRegression
sklearn.metrics
r2_score 
mean_squared_error

